Decoding Table Size Fix
The decoding table was initialised to be too small – it should be long enough to accommodate an ASCII character as an index, i.e. 128. Make it 256 in case an 8-bit char comes in. Note to self, probably should initialise the other elements to some known value. Later... Thanks @rkim for the detective work.